Remove target-specific plug-in projects
To enable Rx Gauntlet to test proposed future builds we needed it to be able to generate new plug-ins at runtime using project file rewriting. Having done that, there was no longer any need to have all the specialized versions of the plug-in project - we now just have a single project.