Rationalise folder structure. Add readme files.
The filesystem layout of this repo had inconsistencies

* it only sometimes matched the Solution Folder structure
* two different conventions were in use for where checks and their supporting projects lived

We now have a Checks folder, with subfolder for each kind of check. The check-type-specific subfolder contains the project that performs the check and any supporting projects it requires.

There is now a top-level readme, and readme files in each of the check type folders.